requirements.txt
Flask==3.0.0
APScheduler==3.10.4
pandas==2.1.4
numpy==1.26.2
statsmodels==0.14.1
fredapi==0.5.1
python-dotenv==1.0.0
gunicorn==21.2.0
